Modified the semantics of max. succ. to enumerate, 0 means unlimited
Previously, 0 meant that the solver should not be used (resulting in 1
successor), and an Optional.empty() meant unlimited successors. However,
handling two special values (0 and unlimited) was complex, and having
the option for 0 solver calls did not really have large benefits.
Therefore, the option for 0 solver calls was removed and instead 0 means
unlimited number of successors to enumerate.