Implement notifyOne part 1
Refactor Transformations to be able to add new stmt.
Added some type enforcing changes too.

Add new stmt: InternalNotifyStmt.
This lets us implement every quirk of this stmt.
See DefaultTransformation::copy(_, Edge) comment for details.
Note that it uses atomic statements, which must be optimized for this case!

Because this depends on XCFA subproject (with the XCFA.Process argument),
XcfaInternalNotifyStmt is added to core subproject, just like XcfaCallStmt.

Added spurious wakeup to XcfaChecker config.
Rewire usage of DefaultTransformation so that it can use the configuration.

Add XcfaStmtVisitorBase which implements XcfaStmtVisitor throwing OperationNotSupported:
New Xcfa stmts are inserted every now and then.
Adding a Stmt needs a change in the XcfaStmtVisitor,
which in turn needs a change in *every class* implementing the interface.
Now every class that used to implement XcfaStmtVisitor now extends XcfaStmtVisitorBase,
so changes in the core does not break every package.

Small fixes and appends to comments.

Note: These features are yet to be tested.