Fix/update XcfaEdgeSplitterTransformation
With this change, the transformation should work as expected, except for
call parameters and return values:

Set names for procedures and processes.
Fix CallStmt referencing the old XCFA procedures.
Make XcfaEdgeSplitterTransformation instance-based (instead of static).
Fix some errors in the code.
Add name of the target location to intermediate edges when splitting to
avoid name collision.