Fix simplifying of fp NaN literal (non)equalities
NaN floating point values should not be equal to any other value. This
behavior is correctly implemented in the FpLitExpr #eq and #neq methods.
To utilize that, manual comparison in simplification is now swapped to
call these methods, resulting in correct behaviour: (= (NaN) (NaN)) for
example now simplifies to false.